Reorganize layout components.
