+++ packages/strum-react/src/components/Text/Text.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Box, BoxProps } from '../Box';
+import { Box, BoxProps } from '../../layouts';
 import * as styles from './Text.css';
 
 export type TextProps = {
