+++ packages/strum-react/src/components/index.ts
@@ -1,11 +1,9 @@
 export * from './Alert';
 export * from './Avatar';
 export * from './Badge';
-export * from './Box';
 export * from './Button';
 export * from './Card';
 export * from './Checkbox';
-export * from './Container';
 export * from './Heading';
 export * from './Icons';
 export * from './Input';
@@ -14,8 +12,6 @@ export * from './Radio';
 export * from './Select';
 export * from './Skeleton';
 export * from './Spinner';
-export * from './Stack';
-export * from './StrumProvider';
 export * from './Text';
 export * from './Tooltip';
 export * from './VisuallyHidden';
