+++ packages/strum-react/src/layouts/Box/Box.docs.mdx
